Acho que já temos basicamente tudo para testar o projeto comos mapas, pedia-os só que fizessem um main onde fizessem só a parte de ler o ficheiro para inicializar a estrutura grid, queeu depois faço aparte do processamento inicial, ou seja, criar as childs iniciais etc... e depois é só correr isto

+++ POO/POO/src/poo/Child.java
@@ -2,7 +2,7 @@ package poo;
 import java.util.*;
 public class Child {
 
-	double comfort;
+	int comfort;
 	ArrayList<Point> path;// Lista de pontos com o caminho percorrido pelo child
 	int length;
 	int cost;
